+++ Changelog.md
@@ -1,3 +1,7 @@
+# New
+
+* [#3](https://github.com/vizor-games/jac/issues/2) Allow reference local values inside string evaluations
+
 # Version 0.0.3
 
 * Allow load configuration from custom locations